backend/requirements.txt
modal>=0.68.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
boto3>=1.28.0
